+++ boto3/resources/base.py
@@ -11,7 +11,12 @@
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
 
+import logging
+
 import boto3
+
+
+logger = logging.getLogger(__name__)
 
 
 class ResourceMeta(object):
@@ -100,6 +105,8 @@ class ServiceResource(object):
                 continue
 
             if name not in self.meta.identifiers:
+                logger.debug('kwargs: %s', kwargs)
+                logger.debug('valid identifiers: %s', self.meta.identifiers)
                 raise ValueError('Unknown keyword argument: {0}'.format(name))
 
             setattr(self, name, value)