+++ tests/unit/resources/test_factory.py
@@ -85,12 +85,33 @@ class TestResourceFactory(BaseTestCase):
         self.assertIn("'handle'", repr(resource))
 
     def test_factory_creates_dangling_resources(self):
+        model = {
+            'has': {
+                'Queue': {
+                    'resource': {
+                        'type': 'Queue',
+                        'identifiers': [
+                            {'target': 'Url', 'source': 'input'}
+                        ]
+                    }
+                },
+                'Message': {
+                    'resource': {
+                        'type': 'Message',
+                        'identifiers': [
+                            {'target': 'QueueUrl', 'source': 'input'},
+                            {'target': 'Handle', 'source': 'input'}
+                        ]
+                    }
+                }
+            }
+        }
         defs = {
             'Queue': {},
             'Message': {}
         }
 
-        TestResource = self.load('test', 'test', {}, defs, None)
+        TestResource = self.load('test', 'test', model, defs, None)
 
         self.assertTrue(hasattr(TestResource, 'Queue'),
             'Missing Queue class from model')
@@ -165,6 +186,18 @@ class TestResourceFactory(BaseTestCase):
             'Object is not an instance of ServiceResource')
 
     def test_dangling_resources_create_resource_instance(self):
+        model = {
+            'has': {
+                'Queue': {
+                    'resource': {
+                        'type': 'Queue',
+                        'identifiers': [
+                            {'target': 'Url', 'source': 'input'}
+                        ]
+                    }
+                }
+            }
+        }
         defs = {
             'Queue': {
                 'identifiers': [
@@ -173,13 +206,25 @@ class TestResourceFactory(BaseTestCase):
             }
         }
 
-        resource = self.load('test', 'test', {}, defs, None)()
+        resource = self.load('test', 'test', model, defs, None)()
         q = resource.Queue('test')
 
         self.assertIsInstance(q, ServiceResource,
             'Dangling resource instance not a ServiceResource')
 
     def test_dangling_resource_create_with_kwarg(self):
+        model = {
+            'has': {
+                'Queue': {
+                    'resource': {
+                        'type': 'Queue',
+                        'identifiers': [
+                            {'target': 'Url', 'source': 'input'}
+                        ]
+                    }
+                }
+            }
+        }
         defs = {
             'Queue': {
                 'identifiers': [
@@ -188,13 +233,25 @@ class TestResourceFactory(BaseTestCase):
             }
         }
 
-        resource = self.load('test', 'test', {}, defs, None)()
+        resource = self.load('test', 'test', model, defs, None)()
         q = resource.Queue(url='test')
 
         self.assertIsInstance(q, ServiceResource,
             'Dangling resource created with kwargs is not a ServiceResource')
 
     def test_dangling_resource_shares_client(self):
+        model = {
+            'has': {
+                'Queue': {
+                    'resource': {
+                        'type': 'Queue',
+                        'identifiers': [
+                            {'target': 'Url', 'source': 'input'}
+                        ]
+                    }
+                }
+            }
+        }
         defs = {
             'Queue': {
                 'identifiers': [
@@ -203,13 +260,25 @@ class TestResourceFactory(BaseTestCase):
             }
         }
 
-        resource = self.load('test', 'test', {}, defs, None)()
+        resource = self.load('test', 'test', model, defs, None)()
         q = resource.Queue('test')
 
         self.assertEqual(resource.meta.client, q.meta.client,
             'Client was not shared to dangling resource instance')
 
     def test_dangling_resource_requires_identifier(self):
+        model = {
+            'has': {
+                'Queue': {
+                    'resource': {
+                        'type': 'Queue',
+                        'identifiers': [
+                            {'target': 'Url', 'source': 'input'}
+                        ]
+                    }
+                }
+            }
+        }
         defs = {
             'Queue': {
                 'identifiers': [
@@ -218,12 +287,24 @@ class TestResourceFactory(BaseTestCase):
             }
         }
 
-        resource = self.load('test', 'test', {}, defs, None)()
+        resource = self.load('test', 'test', model, defs, None)()
 
         with self.assertRaises(ValueError):
             resource.Queue()
 
     def test_dangling_resource_raises_for_unknown_arg(self):
+        model = {
+            'has': {
+                'Queue': {
+                    'resource': {
+                        'type': 'Queue',
+                        'identifiers': [
+                            {'target': 'Url', 'source': 'input'}
+                        ]
+                    }
+                }
+            }
+        }
         defs = {
             'Queue': {
                 'identifiers': [
@@ -232,19 +313,33 @@ class TestResourceFactory(BaseTestCase):
             }
         }
 
-        resource = self.load('test', 'test', {}, defs, None)()
+        resource = self.load('test', 'test', model, defs, None)()
 
         with self.assertRaises(ValueError):
             resource.Queue(url='foo', bar='baz')
 
     def test_dangling_resource_equality(self):
+        model = {
+            'has': {
+                'Queue': {
+                    'resource': {
+                        'type': 'Queue',
+                        'identifiers': [
+                            {'target': 'Url', 'source': 'input'}
+                        ]
+                    }
+                }
+            }
+        }
         defs = {
             'Queue': {
-                'identifiers': [{'name': 'Url'}]
+                'identifiers': [
+                    {'name': 'Url'}
+                ]
             }
         }
 
-        resource = self.load('test', 'test', {}, defs, None)()
+        resource = self.load('test', 'test', model, defs, None)()
 
         q1 = resource.Queue('url')
         q2 = resource.Queue('url')
@@ -252,20 +347,44 @@ class TestResourceFactory(BaseTestCase):
         self.assertEqual(q1, q2)
 
     def test_dangling_resource_inequality(self):
+        model = {
+            'has': {
+                'Queue': {
+                    'resource': {
+                        'type': 'Queue',
+                        'identifiers': [
+                            {'target': 'Url', 'source': 'input'}
+                        ]
+                    }
+                }
+            }
+        }
         defs = {
             'Queue': {
-                'identifiers': [{'name': 'Url'}]
+                'identifiers': [{'name': 'Url'}],
+                'has': {
+                    'Message': {
+                        'resource': {
+                            'type': 'Message',
+                            'identifiers': [
+                                {'target': 'QueueUrl', 'source': 'identifier',
+                                 'name': 'Url'},
+                                {'target': 'Handle', 'source': 'input'}
+                            ]
+                        }
+                    }
+                }
             },
             'Message': {
                 'identifiers': [{'name': 'QueueUrl'}, {'name': 'Handle'}]
             }
         }
 
-        resource = self.load('test', 'test', {}, defs, None)()
+        resource = self.load('test', 'test', model, defs, None)()
 
         q1 = resource.Queue('url')
         q2 = resource.Queue('different')
-        m = resource.Message('url', 'handle')
+        m = q1.Message('handle')
 
         self.assertNotEqual(q1, q2)
         self.assertNotEqual(q1, m)
@@ -299,9 +418,17 @@ class TestResourceFactory(BaseTestCase):
                 'identifiers': [
                     {'name': 'Url'}
                 ],
-                'subResources': {
-                    'resources': ['Message'],
-                    'identifiers': {'Url': 'QueueUrl'}
+                'has': {
+                    'Message': {
+                        'resource': {
+                            'type': 'Message',
+                            'identifiers': [
+                                {'target': 'QueueUrl', 'source': 'identifier',
+                                 'name': 'Url'},
+                                {'target': 'ReceiptHandle', 'source': 'input'}
+                            ]
+                        }
+                    }
                 }
             },
             'Message': {
@@ -500,26 +627,19 @@ class TestResourceFactory(BaseTestCase):
         model = {
             'shape': 'InstanceShape',
             'identifiers': [{'name': 'GroupId'}],
-            'belongsTo': {
+            'has': {
                 'Subnet': {
                     'resource': {
                         'type': 'Subnet',
                         'identifiers': [
-                            {'target': 'Id', 'sourceType': 'dataMember',
-                             'source': 'SubnetId'}
+                            {'target': 'Id', 'source': 'data',
+                             'path': 'SubnetId'}
                         ]
                     }
                 }
             }
         }
         defs = {
-            'Group': {
-                'identifiers': [{'name': 'Id'}],
-                'subResources': {
-                    'identifiers': {'Id': 'GroupId'},
-                    'resources': ['Instance']
-                }
-            },
             'Subnet': {
                 'identifiers': [{'name': 'Id'}]
             }
@@ -529,9 +649,6 @@ class TestResourceFactory(BaseTestCase):
                 'InstanceShape': {
                     'type': 'structure',
                     'members': {
-                        'GroupId': {
-                            'shape': 'String'
-                        },
                         'SubnetId': {
                             'shape': 'String'
                         }
@@ -553,8 +670,6 @@ class TestResourceFactory(BaseTestCase):
                         'Resource should have a subnet reference')
         self.assertIsNone(resource.subnet,
                           'Missing identifier, should return None')
-        self.assertTrue(hasattr(resource, 'group'),
-                        'Resource should have a group reverse ref')
 
         # Load the resource with data to instantiate a reference
         resource.meta.data = {'SubnetId': 'abc123'}
@@ -594,8 +709,8 @@ class TestResourceFactory(BaseTestCase):
                 "Exists": {
                 "waiterName": "BucketExists",
                 "params": [
-                    {"target": "Bucket", "sourceType": "identifier",
-                     "source": "Name"}]
+                    {"target": "Bucket", "source": "identifier",
+                     "name": "Name"}]
                 }
             }
         }
@@ -617,8 +732,8 @@ class TestResourceFactory(BaseTestCase):
                 "Exists": {
                 "waiterName": "BucketExists",
                 "params": [
-                    {"target": "Bucket", "sourceType": "identifier",
-                     "source": "Name"}]
+                    {"target": "Bucket", "source": "identifier",
+                     "name": "Name"}]
                 }
             }
         }